add validations and exception handling

+++ src/main/java/com/readflow/api/service/BookService.java
@@ -1,6 +1,7 @@
 package com.readflow.api.service;
 
 import com.readflow.api.entity.Book;
+import com.readflow.api.exception.global.exceptions.BookNotFoundException;
 import com.readflow.api.repository.BookRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
@@ -19,7 +20,7 @@ public class BookService {
 
     public Book findById(Long id) {
         return bookRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Book not found with id ["+id+"]"));
+                .orElseThrow(() -> new BookNotFoundException("Book not found with id: "+id));
     }
 
     public List<Book> findByTitle(String title){
@@ -37,7 +38,7 @@ public class BookService {
     @Transactional
     public Book update(Long id, Book book) {
         Book existing = bookRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Book not found with id ["+id+"]"));
+                .orElseThrow(() -> new BookNotFoundException("Book not found with id: "+id));
 
         existing.setTitle(book.getTitle());
         existing.setAuthor(book.getAuthor());
@@ -48,6 +49,9 @@ public class BookService {
     }
 
     public void delete(Long id){
+        if (!bookRepository.existsById(id)) {
+            throw new BookNotFoundException("Book not found with id: " + id);
+        }
         bookRepository.deleteById(id);
     }
 }
